requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pyjwt[crypto]==2.10.1
cryptography>=42.0.0
pydantic-settings==2.7.1
httpx==0.28.1
python-dotenv==1.0.1
supabase==2.11.0
python-multipart==0.0.20
